refactor: float -> np.float32

+++ mpm/src/geometries.py
@@ -11,7 +11,7 @@ class Geometry(ABC):
         self.position = np.zeros(shape=(n_particles, 2), dtype=np.float32)
         self.velocity = np.zeros(shape=(n_particles, 2), dtype=np.float32)
         self.phase = np.full(shape=n_particles, fill_value=phase, dtype=int)
-        self.temperature = np.full(shape=n_particles, fill_value=temperature, dtype=float)
+        self.temperature = np.full(shape=n_particles, fill_value=temperature, dtype=np.float32)
         self.frame_threshold = np.zeros(shape=n_particles, dtype=int)
         self.state = np.zeros(shape=n_particles, dtype=int)
